+++ Operaciones/Operaciones.h
@@ -3,4 +3,4 @@
 int Sumar(int s1, int s2);
 int Restar(int s1, int s2);
 int Multiplicar(int s1, int s2);
-float Dividir(int s1, int s2);
+float Dividir(float s1, float s2);
